+++ rt4core/src/main/java/uk/co/farowl/vsj4/stringlib/package-info.java
@@ -1,7 +1,7 @@
 /**
  * The {@code stringlib} package contains classes that support the
- * implementation of built-in types, and relevant to strings, formatting
- * or arrays of integers (that feature heavily in the implementation of
+ * implementation of built-in types, and relevant to strings or arrays
+ * of integers (that feature heavily in the implementation of
  * {@code str}).
  * <p>
  * Classes {@code public} in this package are provisionally intended to
